Cleanup
Conflicts:
	spark-shell/src/main/scala/org/apache/mahout/sparkbindings/shell/Main.scala
